pyproject.toml
[tool.poetry]
name = "monochromist"
version = "1.0.3"
description = "Convert your sketch into contour with transparent background"
authors = ["Bogdan Neterebskii <bog2dan1@gmail.com>"]
license = "MIT License"
repository = "https://github.com/ozzzzz/monochromist"
homepage = "https://github.com/ozzzzz/monochromist"

readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.2.0"
numpy = "^1.20.2"
colour = "^0.1.5"
click = "^7.1.2"
loguru = "^0.5.3"
opencv-python = "^4.5.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
monochromist_opencv = "monochromist.opencv_version.process:process"
monochromist = "monochromist.math_version.process:process"

[tool.black]
line-length = 100
target_version = ["py36"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
